chore(rebase all from autogen): free errors in codes but still unstable yet
