+++ demodapk/__main__.py
@@ -1,4 +1,4 @@
-from autogen import main
+from demodapk.cli import main
 
 if __name__ == "__main__":
     main()
