+++ demodapk/patch.py
@@ -1,0 +1,433 @@
+import os
+import re
+import glob
+import shutil
+from demodapk.utils import msg
+
+
+def extract_package_info(manifest_file):
+    package_name = None
+    package_path = None
+
+    if os.path.isfile(manifest_file):
+        with open(manifest_file, "r", encoding="utf-8") as file:
+            content = file.read()
+
+        # Extract the package name using regex
+        package_match = re.search(r'package="([\w\.]+)"', content)
+        if package_match:
+            package_name = package_match.group(1)
+            package_path = "L" + package_name.replace(".", "/")
+        else:
+            msg.error(f"Package name not found in {manifest_file}.")
+    else:
+        msg.error("AndroidManifest.xml not found.")
+    return package_name, package_path
+
+
+def update_app_name_values(app_name, value_strings):
+    if not os.path.isfile(value_strings):
+        msg.error(f"File not found: {value_strings}")
+        return
+
+    with open(value_strings, "r", encoding="utf-8") as f:
+        content = f.read()
+
+    if '<string name="app_name">' not in content:
+        msg.error("app_name string not found.")
+        return
+
+    new_content = re.sub(
+        r'<string name="app_name">.*?</string>',
+        f'<string name="app_name">{app_name}</string>',
+        content,
+    )
+
+    with open(value_strings, "w", encoding="utf-8") as f:
+        f.write(new_content)
+
+    msg.success(f"Updated AppName to: {app_name}")
+
+
+def update_facebook_app_values(
+    strings_file, fb_app_id, fb_client_token, fb_login_protocol_scheme
+):
+    if os.path.isfile(strings_file):
+        with open(strings_file, "r", encoding="utf-8") as file:
+            content = file.read()
+
+        # Replace values in strings.xml
+        content = re.sub(
+            r'<string name="facebook_app_id">.*?</string>',
+            f'<string name="facebook_app_id">{fb_app_id}</string>',
+            content,
+        )
+        content = re.sub(
+            r'<string name="facebook_client_token">.*?</string>',
+            f'<string name="facebook_client_token">{fb_client_token}</string>',
+            content,
+        )
+        content = re.sub(
+            r'<string name="fb_login_protocol_scheme">.*?</string>',
+            f'<string name="fb_login_protocol_scheme">{fb_login_protocol_scheme}</string>',
+            content,
+        )
+
+        with open(strings_file, "w", encoding="utf-8") as file:
+            file.write(content)
+        msg.success("Updated Facebook App values")
+    else:
+        msg.error(f"File: {strings_file}, does not exists.")
+
+
+def update_files_from_loaded(files_info, apk_dir):
+    """Modify files based on the configuration entry (replace, copy, move within APK)."""
+
+    for operation, files in files_info.items():
+        if operation == "replace":
+            for src, dest in files.items():
+                if isinstance(dest, list):
+                    msg.error(f"Replace does not support multiple destinations: {src}")
+                    continue  # Skip invalid entry
+
+                dest_path = os.path.join(apk_dir, dest)
+
+                if not os.path.exists(src):
+                    msg.error(f"Source file '{src}' not found.")
+                    continue
+
+                os.makedirs(os.path.dirname(dest_path), exist_ok=True)
+
+                try:
+                    shutil.move(src, dest_path)
+                    msg.success(f"Replaced: {src} → {dest_path}")
+                except Exception as e:
+                    msg.error(f"Failed to replace '{src}' → '{dest_path}': {e}")
+
+        elif operation == "copy":
+            for src, dests in files.items():
+                if not isinstance(dests, list):
+                    dests = [dests]  # Ensure it's a list
+
+                if not os.path.exists(src):
+                    msg.error(f"Source file '{src}' not found.")
+                    continue
+
+                for dest in dests:
+                    dest_path = os.path.join(apk_dir, dest)
+                    os.makedirs(os.path.dirname(dest_path), exist_ok=True)
+
+                    try:
+                        shutil.copy2(src, dest_path)
+                        msg.success(f"Copied: {src} → {dest_path}")
+                    except Exception as e:
+                        msg.error(f"Failed to copy '{src}' → '{dest_path}': {e}")
+
+        elif operation == "base_move":
+            for src, dest in files.items():
+                if isinstance(dest, list):
+                    msg.error(f"Move does not support multiple destinations: {src}")
+                    continue  # Skip invalid entry
+
+                src = os.path.join(apk_dir, src)
+                dest_path = os.path.join(apk_dir, dest)
+
+                if not os.path.exists(src):
+                    msg.error(f"Source file '{src}' not found.")
+                    continue
+
+                os.makedirs(os.path.dirname(dest_path), exist_ok=True)
+
+                try:
+                    shutil.move(src, dest_path)
+                    msg.success(f"Moved: {src} → {dest_path}")
+                except Exception as e:
+                    msg.error(f"Failed to move '{src}' → '{dest_path}': {e}")
+
+
+def rename_package_in_manifest(
+    manifest_file, old_package_name, new_package_name, level=0
+):
+    if not os.path.isfile(manifest_file):
+        msg.error("AndroidManifest.xml was not found.")
+        return
+
+    try:
+        with open(manifest_file, "r", encoding="utf-8") as file:
+            content = file.read()
+
+        # Normalize line endings (for Windows compatibility)
+        content = content.replace("\r\n", "\n")
+
+        # Base replacements
+        replacements = [
+            (f'package="{old_package_name}"', f'package="{new_package_name}"'),
+            (
+                f'android:name="{old_package_name}\\.',
+                f'android:name="{new_package_name}.',
+            ),
+            (
+                f'android:authorities="{old_package_name}\\.',
+                f'android:authorities="{new_package_name}.',
+            ),
+        ]
+
+        # Add additional replacements for level == 1
+        if level == 1:
+            replacements.extend(
+                [
+                    (
+                        f'android:taskAffinity="{old_package_name}"',
+                        'android:taskAffinity=""',
+                    ),
+                ]
+            )
+
+        if level == 2:
+            replacements.extend(
+                [
+                    (
+                        f'android:taskAffinity="{old_package_name}"',
+                        f'android:taskAffinity="{new_package_name}"',
+                    ),
+                ]
+            )
+
+        if level == 3:
+            replacements.extend(
+                [
+                    (
+                        f'android:taskAffinity="{old_package_name}"',
+                        f'android:taskAffinity="{new_package_name}"',
+                    ),
+                    (
+                        f'android:host="{old_package_name}"',
+                        f'android:host="{new_package_name}"',
+                    ),
+                    (
+                        f'android:host="cct\\.{old_package_name}"',
+                        f'android:host="cct.{new_package_name}"',
+                    ),
+                ]
+            )
+
+        if level == 4:
+            replacements.extend(
+                [
+                    (
+                        f'android:taskAffinity="{old_package_name}"',
+                        'android:taskAffinity=""',
+                    ),
+                    (
+                        f'android:host="{old_package_name}"',
+                        f'android:host="{new_package_name}"',
+                    ),
+                    (
+                        f'android:host="cct\\.{old_package_name}"',
+                        f'android:host="cct.{new_package_name}"',
+                    ),
+                ]
+            )
+
+        # Perform replacements
+        for pattern, replacement in replacements:
+            content = re.sub(pattern, replacement, content)
+
+        with open(manifest_file, "w", encoding="utf-8") as file:
+            file.write(content)
+
+        msg.success("Updated package name in AndroidManifest.xml")
+
+    except FileNotFoundError:
+        msg.error(f"The manifest file '{manifest_file}' was not found.")
+    except IOError as e:
+        msg.error(f"Error reading or writing the manifest file: {e}")
+    except re.error as e:
+        msg.error(f"Regular expression error: {e}")
+
+
+def update_smali_path_package(smali_dir, old_package_path, new_package_path):
+    for root, _, files in os.walk(smali_dir):
+        for file in files:
+            file_path = os.path.join(root, file)
+            if file.endswith(".smali"):
+                with open(file_path, "r", encoding="utf-8") as f:
+                    content = f.read()
+
+                # Replace old package name with new one
+                new_content = content.replace(old_package_path, new_package_path)
+
+                with open(file_path, "w", encoding="utf-8") as f:
+                    f.write(new_content)
+
+    msg.success("Updated package name in smali files.")
+
+
+def rename_package_in_resources(resources_dir, old_package_name, new_package_name):
+    try:
+        # Check if resources directory exists
+        if not os.path.isdir(resources_dir):
+            raise FileNotFoundError(
+                f"The resources directory '{resources_dir}' does not exist."
+            )
+
+        updated_any = False  # Track if any file was updated
+        valid_file_extensions = (
+            ".xml",
+            ".json",
+            ".txt",
+        )  # Define file types to process
+
+        for root, _, files in os.walk(resources_dir):
+            for file in files:
+                if not file.endswith(valid_file_extensions):
+                    continue  # Skip irrelevant file types
+
+                file_path = os.path.join(root, file)
+
+                try:
+                    with open(file_path, "r", encoding="utf-8") as f:
+                        content = f.read()
+
+                    # Check if the old package name exists in the file
+                    if f'"{old_package_name}"' in content:
+                        new_content = content.replace(
+                            f'"{old_package_name}"', f'"{new_package_name}"'
+                        )
+                        with open(file_path, "w", encoding="utf-8") as f:
+                            f.write(new_content)
+                        updated_any = True
+
+                except UnicodeDecodeError:
+                    msg.warning(f"File {file_path} is not UTF-8 encoded, skipping.")
+                except (OSError, IOError) as e:
+                    msg.error(f"Failed to process file: {file_path}. Error: {e}")
+
+        if updated_any:
+            msg.success("Updated package name in resource files.")
+        else:
+            msg.info("No files with the specified package name were found to update.")
+
+    except FileNotFoundError as fnf_error:
+        msg.error(str(fnf_error))
+    except Exception as e:
+        msg.error(f"An unexpected error occurred: {e}")
+
+
+def update_smali_directory(smali_dir, old_package_path, new_package_path):
+    # Use glob to find all directories in the smali_dir that match the old
+    old_package_pattern = os.path.join(
+        smali_dir, "**", old_package_path.strip("L")
+    )  # Use '**' to search in subdirectories
+    # Recursively find all matching directories
+    old_dirs = glob.glob(old_package_pattern, recursive=True)
+
+    renamed = False  # Track if any directory was renamed
+
+    for old_dir in old_dirs:
+        # Create the new directory path based on the found old directory
+        new_dir = old_dir.replace(
+            old_package_path.strip("L"), new_package_path.strip("L")
+        )
+
+        if os.path.isdir(old_dir):
+            # Rename the old directory to the new directory
+            os.rename(old_dir, new_dir)
+            msg.success(f"Renamed directory {old_dir} to {new_dir}.")
+            renamed = True
+        else:
+            msg.info(f"Directory {old_dir} does not exist. Skipping renaming.")
+
+    if not renamed:
+        msg.info(
+            f"No directories matching {old_package_path} were found. Skipping renaming."
+        )
+
+
+def update_application_id_in_smali(smali_dir, old_package_name, new_package_name):
+    try:
+        # Check if smali directory exists
+        if not os.path.isdir(smali_dir):
+            raise FileNotFoundError(
+                f"The smali directory '{smali_dir}' does not exist."
+            )
+
+        buildconfig_found = False
+        updated_any = False
+
+        for root, _, files in os.walk(smali_dir):
+            for file in files:
+                if file.endswith("BuildConfig.smali"):
+                    buildconfig_found = True
+                    file_path = os.path.join(root, file)
+                    try:
+                        with open(file_path, "r", encoding="utf-8") as f:
+                            content = f.read()
+
+                        # Check if old APPLICATION_ID exists in the file
+                        if (
+                            f' APPLICATION_ID:Ljava/lang/String; = "{old_package_name}"'
+                            in content
+                        ):
+                            # Replace APPLICATION_ID
+                            new_content = content.replace(
+                                f' APPLICATION_ID:Ljava/lang/String; = "{old_package_name}"',
+                                f' APPLICATION_ID:Ljava/lang/String; = "{new_package_name}"',
+                            )
+                            with open(file_path, "w", encoding="utf-8") as f:
+                                f.write(new_content)
+                            updated_any = True
+
+                    except (OSError, IOError) as e:
+                        msg.error(f"Failed to update {file_path}: {e}")
+
+        if not buildconfig_found:
+            raise FileNotFoundError(
+                "No BuildConfig.smali files found in the provided smali directory."
+            )
+        if not updated_any:
+            raise ValueError(
+                "No BuildConfig.smali file contained the specified old APPLICATION_ID."
+            )
+    except FileNotFoundError as fnf_error:
+        msg.error(str(fnf_error))
+    except ValueError as val_error:
+        msg.error(str(val_error))
+    except Exception as e:
+        msg.error(f"An unexpected error occurred: {e}")
+    else:
+        msg.success("Updated APPLICATION_ID in smali files.")
+
+
+def remove_metadata_from_manifest(manifest_file, config_file):
+    # Filter out any empty strings in metadata_to_remove
+    metadata_to_remove = [
+        meta for meta in config_file if isinstance(meta, str) and meta.strip()
+    ]
+
+    # If metadata_to_remove is empty or only contained empty strings, skip removal
+    if not metadata_to_remove:
+        # msg.info("No valid metadata entries specified for removal in configuration file.")
+        return
+
+    if os.path.isfile(manifest_file):
+        with open(manifest_file, "r", encoding="utf-8") as file:
+            lines = file.readlines()
+
+        # Filter out lines that match the metadata
+        filtered_lines = []
+        skip = False
+        for line in lines:
+            if any(meta in line for meta in metadata_to_remove):
+                skip = True  # Skip this line and the next
+            elif skip and "/>" in line:  # End of the meta-data entry
+                skip = False  # Stop skipping after closing tag
+            else:
+                filtered_lines.append(line)
+
+        # Write the filtered lines back to the manifest file
+        with open(manifest_file, "w", encoding="utf-8") as file:
+            file.writelines(filtered_lines)
+        msg.success("Specific metadata removed from manifest.")
+    else:
+        msg.warning(f"File '{manifest_file}' does not exist.")
